fix: 내 사전 의견 삭제 API 삭제 방식 수정
